+++ packages/card/src/index.tsx
@@ -46,7 +46,16 @@ export type ProCardProps = {
    */
   title?: React.ReactNode;
   /**
+   * 副标题
+   */
+  subTitle?: React.ReactNode;
+  /**
    * 标题说明
+   */
+  tooltip?: string;
+
+  /**
+   * @deprecated 你可以使用 tooltip，这个更改是为了与 antd 统一
    */
   tip?: string;
   /**
@@ -57,6 +66,10 @@ export type ProCardProps = {
    * 布局，center 代表垂直居中
    */
   layout?: 'default' | 'center';
+  /**
+   * 卡片类型
+   */
+  type?: 'inner';
   /**
    * 加载中
    */
@@ -106,12 +119,14 @@ const ProCard: ProCardType = (props) => {
     bodyStyle = {},
     headStyle = {},
     title,
+    subTitle,
     extra,
     tip,
     layout,
     loading,
     colSpan,
     gutter = 0,
+    tooltip,
     split,
     headerBordered = false,
     bordered = false,
@@ -121,6 +136,7 @@ const ProCard: ProCardType = (props) => {
     collapsible = false,
     defaultCollapsed = false,
     onCollapse,
+    type,
   } = props;
 
   const screens = useBreakpoint();
@@ -248,6 +264,7 @@ const ProCard: ProCardType = (props) => {
           [`${prefixCls}-loading`]: loading,
           [`${prefixCls}-split`]: split === 'vertical' || split === 'horizontal',
           [`${prefixCls}-ghost`]: ghost,
+          [`${prefixCls}-type-${type}`]: type,
         });
 
         const headerCls = classNames(`${prefixCls}-header`, {
@@ -287,7 +304,7 @@ const ProCard: ProCardType = (props) => {
             {(title || extra || collapsibleButton) && (
               <div className={headerCls} style={headStyle}>
                 <div className={`${prefixCls}-title`}>
-                  <LabelIconTip label={title} tip={tip} />
+                  <LabelIconTip label={title} tooltip={tooltip || tip} subTitle={subTitle} />
                   {collapsibleButton}
                 </div>
                 <div className={`${prefixCls}-extra`}>{extra}</div>