+++ packages/form/src/demos/layout-base.tsx
@@ -59,7 +59,7 @@ export default () => {
                 <ProFormText
                   name="name"
                   label="签约客户名称"
-                  tip="最长为 24 位"
+                  tooltip="最长为 24 位"
                   placeholder="请输入名称"
                 />
                 <ProFormText name="company" label="我方公司名称" placeholder="请输入名称" />